jd01-01 jd01-02 jd01-03 jd01-04 jd01-05 jd01-06

+++ src/by/it/seledtsova/jd01_07/Vector.java
@@ -3,7 +3,7 @@ package by.it.seledtsova.jd01_07;
 public class Vector extends Var{
     private  final  int[] value;
 
-    public Vector(int [] vector) {
+    public Vector(double[] vector) {
         this.value = new int[0];
     }
